Extract all Haskell rules to their own modules
Following the trend, we split out everything for Haskell rules.

There is quite a bit of boilerplate happening here.
That seems fine though since it means rules become more accurate.
That accuracy means that shake shouldn't have to do so much work.
